+++ src/dotNet/Animations.cs
@@ -133,13 +133,17 @@ namespace DesktopPet
                 /// </summary>
             VERTICAL    = 0x08,
                 /// <summary>
+                /// Vertical screen, window side and pet side borders - next animation will be executed only if pet hits any vertical border
+                /// </summary>
+            VERTICAL_ = 0x0A,
+                /// <summary>
                 /// Pet-side borders - next animation will be executed only if pet hits the left or right of another pet
                 /// </summary>
-            PETSIDE = 0x0A,
+            PETSIDE = 0x0C,
                 /// <summary>
                 /// Window-side borders - next animation will be executed only if pet hits the left or right of a window
                 /// </summary>
-            WINDOWSIDE = 0x0C
+            WINDOWSIDE = 0x0E,
         }
             /// <summary>
             /// ID of the next animation to play
@@ -800,11 +804,13 @@ namespace DesktopPet
             /// <returns>ID of the next animation to play. -1 if there is no animation.</returns>
         private int SetNextGeneralAnimation(List<TNextAnimation> list, TNextAnimation.TOnly where)
         {
-            //Console.WriteLine(where);
-
             int iDefaultID = -1;
             if (list.Count > 0)     // Find the next animation only if there is at least 1 animation in the list
             {
+                // Determine if horizontal+ or vertical+ animations should also be considered
+                bool isHorizontal_ = (where == TNextAnimation.TOnly.WINDOW || where == TNextAnimation.TOnly.TASKBAR || where == TNextAnimation.TOnly.HORIZONTAL) ? true : false;
+                bool isVertical_ = (where == TNextAnimation.TOnly.VERTICAL || where == TNextAnimation.TOnly.WINDOWSIDE || where == TNextAnimation.TOnly.PETSIDE) ? true : false;
+
                 int iVal;
                 int iSum = 0;
                 int iRandMax = 0;
@@ -812,7 +818,13 @@ namespace DesktopPet
                 foreach (TNextAnimation anim in list)
                 {
                     // Skip if this animation has the wrong "only" value
-                    if (anim.only != TNextAnimation.TOnly.NONE && anim.only != where) continue;
+                    if (anim.only != TNextAnimation.TOnly.NONE
+                        && !(isHorizontal_ && anim.only == TNextAnimation.TOnly.HORIZONTAL_)
+                        && !(isVertical_ && anim.only == TNextAnimation.TOnly.VERTICAL_)
+                        && anim.only != where)
+                    {
+                        continue;
+                    }
 
                     //Console.WriteLine("anim: " + SheepAnimations[anim.ID].Name + " | anim.only: " + anim.only + " | where: " + where + " | anim.only & where: " + (anim.only & where));
 
@@ -823,7 +835,14 @@ namespace DesktopPet
                 iVal = rand.Next(1, iRandMax+1);
                 foreach (TNextAnimation anim in list)
                 {
-                    if (anim.only != TNextAnimation.TOnly.NONE && anim.only != where) continue;
+                    // Skip if this animation has the wrong "only" value
+                    if (anim.only != TNextAnimation.TOnly.NONE
+                        && !(isHorizontal_ && anim.only == TNextAnimation.TOnly.HORIZONTAL_)
+                        && !(isVertical_ && anim.only == TNextAnimation.TOnly.VERTICAL_)
+                        && anim.only != where)
+                    {
+                        continue;
+                    }
 
                     iSum += anim.Probability;
                     if (iSum >= iVal)