+++ src/dotNet/Xml.cs
@@ -260,6 +260,7 @@ namespace DesktopPet
                             case "horizontal": where = TNextAnimation.TOnly.HORIZONTAL; break;
                             case "horizontal+": where = TNextAnimation.TOnly.HORIZONTAL_; break;
                             case "vertical": where = TNextAnimation.TOnly.VERTICAL; break;
+                            case "vertical+": where = TNextAnimation.TOnly.VERTICAL_; break;
                             case "petSide": where = TNextAnimation.TOnly.PETSIDE; break;
                             case "windowSide": where = TNextAnimation.TOnly.WINDOWSIDE; break;
                             default: where = TNextAnimation.TOnly.NONE; break;
@@ -287,6 +288,7 @@ namespace DesktopPet
                             case "horizontal": where = TNextAnimation.TOnly.HORIZONTAL; break;
                             case "horizontal+": where = TNextAnimation.TOnly.HORIZONTAL_; break;
                             case "vertical": where = TNextAnimation.TOnly.VERTICAL; break;
+                            case "vertical+": where = TNextAnimation.TOnly.VERTICAL_; break;
                             case "petSide": where = TNextAnimation.TOnly.PETSIDE; break;
                             case "windowSide": where = TNextAnimation.TOnly.WINDOWSIDE; break;
                             default: where = TNextAnimation.TOnly.NONE; break;
@@ -314,6 +316,7 @@ namespace DesktopPet
                             case "horizontal": where = TNextAnimation.TOnly.HORIZONTAL; break;
                             case "horizontal+": where = TNextAnimation.TOnly.HORIZONTAL_; break;
                             case "vertical": where = TNextAnimation.TOnly.VERTICAL; break;
+                            case "vertical+": where = TNextAnimation.TOnly.VERTICAL_; break;
                             case "petSide": where = TNextAnimation.TOnly.PETSIDE; break;
                             case "windowSide": where = TNextAnimation.TOnly.WINDOWSIDE; break;
                             default: where = TNextAnimation.TOnly.NONE; break;